AB feat: some bugs fixed in cart nd orders

+++ Menro.Application/Features/Orders/DTOs/MonthlySalesDto.cs
@@ -4,6 +4,6 @@
     {
         public int Month { get; set; }
         public string MonthName { get; set; } = "";
-        public decimal TotalAmount { get; set; }
+        public int TotalAmount { get; set; }
     }
 }
